requirements.prod.txt
# Base Django requirements
-r requirements.txt

# Production-specific requirements

# WSGI server
gunicorn==21.2.0

# Database drivers with connection pooling
psycopg2-binary==2.9.9
django-db-connection-pool==1.0.2

# Caching and sessions
django-redis==5.4.0
redis==5.0.1

# Environment variables and configuration management
django-environ==0.11.2

# Static and media file storage
django-storages==1.14.2
boto3==1.34.29

# Security
django-csp==3.7
django-cors-headers==4.3.1
django-secure==1.0.2

# Monitoring and metrics
sentry-sdk==1.40.0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Distributed task queue
celery==5.5.2
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Performance and profiling
pyinstrument==4.6.1
django-debug-toolbar==4.2.0

# Testing and quality
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Content security policy
django-csp==3.7
